// src/app/interfaces/pokemon-data.interface.ts
export interface PokemonData {
  abilities:                Ability[];
  base_experience:          number;
  cries:                    Cries;
  forms:                    UrlInfo[];
  game_indices:             GameIndex[];
  height:                   number;
  held_items:               HeldItem[];
  id:                       number;
  is_default:               boolean;
  location_area_encounters: string;
  moves:                    MoveData[];
  name:                     string;
  order:                    number;
  past_abilities:           PastAbility[];
  past_types:               any[];
  species:                  UrlInfo;
  sprites:                  Sprites;
  stats:                    Stat[];
  types:                    PokemonTypeUrl[];
  weight:                   number;
}

export interface Ability {
  ability:   UrlInfo | null;
  is_hidden: boolean;
  slot:      number;
}

export interface UrlInfo {
  name: string;
  url:  string;
}

export interface Cries {
  latest: string;
  legacy: string;
}

export interface GameIndex {
  game_index: number;
  version:    UrlInfo;
}

export interface HeldItem {
  item:            UrlInfo;
  version_details: VersionDetail[];
}

export interface VersionDetail {
  rarity:  number;
  version: UrlInfo;
}

export interface MoveData {
  move:                  UrlInfo;
  version_group_details: VersionGroupDetail[];
}

export interface VersionGroupDetail {
  level_learned_at:  number;
  move_learn_method: UrlInfo;
  order:             null;
  version_group:     UrlInfo;
}

export interface PastAbility {
  abilities:  Ability[];
  generation: UrlInfo;
}

export interface GenerationV {
  "black-white": Sprites;
}

export interface GenerationIv {
  "diamond-pearl":        Sprites;
  "heartgold-soulsilver": Sprites;
  platinum:               Sprites;
}

export interface Versions {
  "generation-i":    GenerationI;
  "generation-ii":   GenerationIi;
  "generation-iii":  GenerationIii;
  "generation-iv":   GenerationIv;
  "generation-v":    GenerationV;
  "generation-vi":   { [key: string]: Home };
  "generation-vii":  GenerationVii;
  "generation-viii": GenerationViii;
}

export interface Other {
  dream_world:        DreamWorld;
  home:               Home;
  "official-artwork": OfficialArtwork;
  showdown:           Sprites;
}

export interface Sprites {
  back_default:       string;
  back_female:        null;
  back_shiny:         string;
  back_shiny_female:  null;
  front_default:      string;
  front_female:       null;
  front_shiny:        string;
  front_shiny_female: null;
  other?:             Other;
  versions:          Versions;
  animated:          Sprites;
}
// ...
export interface Stat {
  base_stat: number;
  effort:    number;
  stat:      UrlInfo;
}

export interface PokemonTypeUrl {
  slot: number;
  type: UrlInfo;
}